+++ MatrixSDK/MXSession.h
@@ -1451,12 +1451,14 @@ typedef void (^MXOnBackgroundSyncFail)(NSError *error);
 /**
  Decrypt an event and update its data.
 
+ @warning This method is deprecated, use -[MXSession decryptEvents:inTimeline:onComplete:] instead.
+ 
  @param event the event to decrypt.
  @param timeline the id of the timeline where the event is decrypted. It is used
         to prevent replay attack.
  @return YES if decryption is successful.
  */
-- (BOOL)decryptEvent:(MXEvent*)event inTimeline:(NSString*)timeline;
+- (BOOL)decryptEvent:(MXEvent*)event inTimeline:(NSString*)timeline  __attribute__((deprecated("use -[MXSession decryptEvents:inTimeline:onComplete:] instead")));
 
 /**
  Decrypt events asynchronously and update their data.